combine the two useSocket hooks
